requirements.txt
wheel
Django == 2.2.28
django-admin-list-filter-dropdown
django-crum
django-fontawesome-5
django-jsonfield
django-mptt
django-mptt-admin
django-select2
django-stdimage
django-widget-tweaks
django-dotenv
mysqlclient
pandas
Pillow
redis
celery
django-celery-beat
django-object-actions
businesstimedelta
admin-totals
django-bootstrap-modal-forms

# Rest API
djangorestframework
django-cors-headers


# Debuging tools
django-silk         # https://github.com/jazzband/django-silk
sentry-sdk          # https://docs.sentry.io/platforms/python/
